# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ERP MAFA Store — API
  version: "0.1.0"
  description: >
    Especificação de API gerada a partir dos controllers do Nest.js
    (apps/api). Ver convenções em docs/architecture/overview.md, seção 10.
servers:
  - url: /api/v1
...
paths:
...
  /purchasing/suppliers:
    post:
      tags: [purchasing]
      summary: Cria fornecedor (requer permission:manage_purchasing)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                document: { type: string, nullable: true }
                contactName: { type: string, nullable: true }
                phone: { type: string, nullable: true }
                whatsapp: { type: string, nullable: true }
                email: { type: string, format: email, nullable: true }
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }
    get:
      tags: [purchasing]
      summary: >
        Lista fornecedores (requer permission:view_purchasing). Sem `page`, array completo
        (retrocompatível — usado pelo seletor de fornecedor em Compras). Com `page`, formato
        `{items,total,page,pageSize}`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: "#/components/schemas/Supplier" }
                  - $ref: "#/components/schemas/PaginatedSuppliers"
  /purchasing/suppliers/{id}:
    get:
      tags: [purchasing]
      summary: Consulta ficha e indicadores do fornecedor (requer permission:view_purchasing)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Fornecedor com resumo comercial e compras recentes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
    patch:
      tags: [purchasing]
      summary: Edita campos cadastrais do fornecedor (requer permission:manage_purchasing)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                document: { type: string, nullable: true }
                contactName: { type: string, nullable: true }
                phone: { type: string, nullable: true }
                whatsapp: { type: string, nullable: true }
                email: { type: string, format: email, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }
  /purchasing/suppliers/{id}/deactivate:
    patch:
      tags: [purchasing]
      summary: Inativa o fornecedor (nunca deleta)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }
  /purchasing/suppliers/{id}/reactivate:
    patch:
      tags: [purchasing]
      summary: Reativa o fornecedor
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }
  /purchasing/purchases:
    post:
      tags: [purchasing]
      summary: Cria uma compra em rascunho com itens (requer permission:manage_purchasing) — RN 10.6.1, não altera estoque
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                supplierId: { type: string, format: uuid }
                currency: { type: string, default: BRL }
                exchangeRate: { type: number }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productVariantId, quantity, unitCostOriginCurrency]
                    properties:
                      productVariantId: { type: string, format: uuid }
                      quantity: { type: number }
                      unitCostOriginCurrency: { type: number }
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Purchase" }
    get:
      tags: [purchasing]
      summary: Lista compras (requer permission:view_purchasing)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Purchase" }
  /purchasing/purchases/{id}:
    get:
      tags: [purchasing]
      summary: Detalhe da compra, com itens e recebimentos (requer permission:view_purchasing)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Purchase" }
        "404":
          description: Compra não encontrada nesta empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      tags: [purchasing]
      summary: Substitui os dados e itens de uma compra em rascunho (requer permission:manage_purchasing)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                supplierId: { type: string, format: uuid }
                currency: { type: string, default: BRL }
                exchangeRate: { type: number }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productVariantId, quantity, unitCostOriginCurrency]
                    properties:
                      productVariantId: { type: string, format: uuid }
                      quantity: { type: number }
                      unitCostOriginCurrency: { type: number }
      responses:
        "200":
          description: Rascunho atualizado; itens anteriores preservados como inativos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Purchase" }
        "409":
          description: Compra não está em rascunho
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /purchasing/purchases/{id}/order:
    post:
      tags: [purchasing]
      summary: draft → ordered (RN 10.6.2, ainda não altera estoque)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Purchase" }
        "409":
          description: Compra não está em rascunho
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /purchasing/purchases/{id}/receive:
    post:
      tags: [purchasing]
      summary: >
        Recebimento total ou parcial (RN 10.6.3/10.6.4/10.6.5). Rateia custos
        adicionais proporcionalmente ao valor dos itens e delega a baixa/entrada
        de estoque ao InventoryService.receiveGoods() na mesma transação
        (TA-ARCH-003) — recalcula o custo médio móvel do produto (RN 11.4).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [purchaseItemId, quantityReceived]
                    properties:
                      purchaseItemId: { type: string, format: uuid }
                      quantityReceived: { type: number }
                additionalCosts:
                  type: array
                  items:
                    type: object
                    required: [type, amount]
                    properties:
                      type: { type: string, example: freight }
                      amount: { type: number }
      responses:
        "201":
          description: Recebido (compra atualizada, com status partially_received ou received)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Purchase" }
        "400":
          description: Item inválido ou quantidade recebida maior que a comprada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Compra não está pedida/parcialmente recebida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /purchasing/purchases/{id}/cancel:
    post:
      tags: [purchasing]
      summary: >
        Cancela a compra — só permitido em draft/ordered. Cancelar uma compra
        já recebida exige estorno de estoque/custo (RN 10.6.6), não
        implementado nesta fase — retorna 409 explícito.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "201":
          description: Cancelada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Purchase" }
        "409":
          description: Compra já recebida ou já cancelada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        TA-API-002 — chave opcional fornecida pelo cliente para evitar duplicar o efeito de um
        comando crítico em caso de reenvio (ex.: timeout de rede). Repetir a mesma chave com o
        mesmo payload devolve a resposta original; com payload diferente é tratado como uma nova
        operação, não como conflito.
      schema: { type: string, maxLength: 255 }
    SearchQuery:
      name: q
      in: query
      required: false
      description: Busca por nome (contains, case-insensitive).
      schema: { type: string }
    PageQuery:
      name: page
      in: query
      required: false
      description: >
        Presença deste parâmetro é o que ativa a paginação — sem ele a resposta é o array
        completo de sempre (retrocompatível com os seletores de produto/fornecedor/cliente
        usados em Compras/Vendas).
      schema: { type: integer, minimum: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
...
    Supplier:
      type: object
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        name: { type: string }
        document: { type: string, nullable: true }
        contactName: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        whatsapp: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        status: { type: string, enum: [active, inactive] }
    SupplierDetail:
      allOf:
        - $ref: "#/components/schemas/Supplier"
        - type: object
          properties:
            summary:
              type: object
              properties:
                purchasesCount: { type: integer }
                totalPurchased: { type: string }
                averagePurchase: { type: string }
                lastPurchaseAt: { type: string, format: date-time, nullable: true }
                productsSupplied: { type: integer }
                outstandingBalance: { type: string }
            recentPurchases:
              type: array
              maxItems: 5
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
                  currency: { type: string }
                  total: { type: string }
                  createdAt: { type: string, format: date-time }
                  itemsCount: { type: integer }
    PaginatedSuppliers:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Supplier" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    Purchase:
      type: object
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        supplierId: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [draft, ordered, partially_received, received, cancelled] }
        currency: { type: string }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              productVariantId: { type: string, format: uuid }
              quantity: { type: string }
              unitCostOriginCurrency: { type: string }
              quantityReceived: { type: string }
        receipts:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              receivedAt: { type: string, format: date-time }
              items:
                type: array
                items:
                  type: object
                  properties:
                    purchaseItemId: { type: string, format: uuid }
                    quantityReceived: { type: string }
                    unitCostFinal: { type: string }
...
    ErrorResponse:
      type: object
      description: Formato padrão TA-API-001 (docs/architecture/overview.md, secao 10.2)
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            field: { type: string }
            details: { type: object }